backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.10.1
flask-redis==0.4.0
Flask-Session==0.8.0
Flask_Limit==2.0.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
openai==1.64.0
ordered-set==4.1.0
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
openai
requests
wikipedia  # For Wikipedia/Wikidata
googlemaps  # For Google Maps API
noaa-sdk  # For NOAA data (install or find equivalent)
gebco  # For GEBCO bathymetric data (custom or find equivalent)
nasa-api  # For NASA Earth Observations (custom or find equivalent)
usgs  # For USGS earthquake data (custom or find equivalent)
python-dotenv
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.10.1
flask-redis==0.4.0
Flask-Session==0.8.0
Flask_Limit==2.0.5
googlemaps==4.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
openai==1.64.0
ordered-set==4.1.0
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
